Module Travesty

module Bi_mappable : sig ... end

Mapping for containers with two element types.

module Bi_mappable_types : sig ... end

The main groups of signatures provided by this module are:

module Bi_traversable : sig ... end

Signatures and functors for containers and data structures that can be mapped across in two different ways with a monadic side-effect.

module Bi_traversable_types : sig ... end

The main groups of signatures provided by this module are:

module Container_exts : sig ... end

Extensions for containers.

module Container_exts_types : sig ... end

Extension signatures for containers.

module Filter_mappable : sig ... end

Simultaneous mapping and filtering.

module Filter_mappable_types : sig ... end

Signatures for filter-mapping.

module Generic_types : sig ... end

Generic type signatures.

module Mappable : sig ... end

Non-monadic mapping.

module Mappable_types : sig ... end

Signatures for (non-monadic) mapping.

module Monad_exts : sig ... end

Generic monad extensions.

module Monad_exts_types : sig ... end

module Monad_transform : sig ... end

Monad transformer signatures.

module State : sig ... end

Haskell-style state monads.

module State_transform : sig ... end

Haskell-style state transformers.

module State_transform_types : sig ... end

module State_types : sig ... end

module Traversable : sig ... end

Signatures and functors for containers and data structures that can be mapped across with a monadic side-effect.

module Traversable_types : sig ... end

The main groups of signatures provided by this module are: