Module Bi_mappable.Chain_Map1_Bi1_right

Chain_Map1_Bi1_right (RMap) (Bi) composes an inner arity-1 map LMap on the right of an arity-1 bi-map Bi with floating right type.

Parameters

module RMap : Mappable_types.S1

module Bi : Bi_mappable_types.Basic1_right

Signature

include Generic_types.Bi_right with type 'r t = 'r RMap.t Bi.t with type left = Bi.left

type 'r t = 'r RMap.t Bi.t

Partially fixed type of containers.

type left = Bi.left

Fixed type of left elements.

include Bi_mappable_types.Generic with type ('l, 'r) t := 'r t and type 'l left := left and type 'r right := 'r

include Bi_mappable_types.Basic_generic with type ('l, 'r) t := 'r t with type 'l left := left with type 'r right := 'r

include Generic_types.Bi_generic with type ('l, 'r) t := 'r t with type 'l left := left with type 'r right := 'r

val bi_map : 'r1 t -> left:(left -> left) -> right:('r1 -> 'r2) -> 'r2 t

bi_map c ~left ~right maps left over every 'l1 left, and right over every 'r1 right, in c.

val map_left : 'r t -> f:(left -> left) -> 'r t

map_left c ~f maps f over the left type of c only.

val map_right : 'r1 t -> f:('r1 -> 'r2) -> 'r2 t

map_right c ~f maps f over the right type of c only.