Module Bi_mappable.Chain_Map1_Bi2

Chain_Map1_Bi2 (LMap) (RMap) (Bi) composes an inner arity-1 map LMap on the left, and another such map RMap on the right, of an arity-2 bi-traversal Bi.

Parameters

module LMap : Mappable_types.S1

module RMap : Mappable_types.S1

module Bi : Bi_mappable_types.Basic2

Signature

Type of containers.

include Base.T2 with type ('l, 'r) t = ( 'l LMap.t, 'r RMap.t ) Bi.t

type ('l, 'r) t = ( 'l LMap.t, 'r RMap.t ) Bi.t

include Bi_mappable_types.Generic with type ('l, 'r) t := ( 'l, 'r ) t and type 'l left := 'l and type 'r right := 'r

include Bi_mappable_types.Basic_generic with type ('l, 'r) t := ( 'l, 'r ) t with type 'l left := 'l with type 'r right := 'r

include Generic_types.Bi_generic with type ('l, 'r) t := ( 'l, 'r ) t with type 'l left := 'l with type 'r right := 'r

val bi_map : ( 'l1, 'r1 ) t -> left:( 'l1 -> 'l2 ) -> right:( 'r1 -> 'r2 ) -> ( 'l2, 'r2 ) t

bi_map c ~left ~right maps left over every 'l1 left, and right over every 'r1 right, in c.

val map_left : ( 'l1, 'r ) t -> f:( 'l1 -> 'l2 ) -> ( 'l2, 'r ) t

map_left c ~f maps f over the left type of c only.

val map_right : ( 'l, 'r1 ) t -> f:( 'r1 -> 'r2 ) -> ( 'l, 'r2 ) t

map_right c ~f maps f over the right type of c only.