Module Bi_mappable.Fix2_both

Fix2_both (S) (Left) (Right) fixes the types of S to Left and Right, making it an S0.

Parameters

module I : Bi_mappable_types.Basic2

module Left : Base.T

module Right : Base.T

Signature

include Generic_types.Bi0 with type t = (Left.t, Right.t) I.t with type left = Left.t with type right = Right.t

type t = (Left.t, Right.t) I.t

Fixed type of containers.

type left = Left.t

Fixed type of left elements.

type right = Right.t

Fixed type of right elements.

include Bi_mappable_types.Generic with type ('l, 'r) t := t and type 'l left := left and type 'r right := right

include Bi_mappable_types.Basic_generic with type ('l, 'r) t := t with type 'l left := left with type 'r right := right

include Generic_types.Bi_generic with type ('l, 'r) t := t with type 'l left := left with type 'r right := right

val bi_map : t -> left:(left -> left) -> right:(right -> right) -> t

bi_map c ~left ~right maps left over every 'l1 left, and right over every 'r1 right, in c.

val map_left : t -> f:(left -> left) -> t

map_left c ~f maps f over the left type of c only.

val map_right : t -> f:(right -> right) -> t

map_right c ~f maps f over the right type of c only.