Module Bi_mappable.Fix2_left

Fix2_left (I) (Left) fixes the left type of I to Left, making it an S1_right.

Parameters

I : Bi_mappable_types.Basic2

Left : Base.T

Signature

include Generic_types.Bi_right

type 'r t

Partially fixed type of containers.

type left

Fixed type of left elements.

include Bi_mappable_types.Generic with type ('l, 'r) t := 'r t and type 'l left := left and type 'r right := 'r

include Bi_mappable_types.Basic_generic

include Generic_types.Bi_generic

type ('l, 'r) t

Generic container type.

type 'l left

Generic left type.

type 'r right

Generic right type.

val bi_map : ('l1, 'r1) t -> left:('l1 left -> 'l2 left) -> right:('r1 right -> 'r2 right) -> ('l2, 'r2) t

bi_map c ~left ~right maps left over every 'l1 left, and right over every 'r1 right, in c.

val map_left : ('l1, 'r) t -> f:('l1 left -> 'l2 left) -> ('l2, 'r) t

map_left c ~f maps f over the left type of c only.

val map_right : ('l, 'r1) t -> f:('r1 right -> 'r2 right) -> ('l, 'r2) t

map_right c ~f maps f over the right type of c only.