Module Bi_mappable.Make0

Make0 implements S0 for an arity-0 bi-mappable container.

Parameters

module I : Bi_mappable_types.Basic0

Signature

include Generic_types.Bi0 with type t = I.t with type left = I.left with type right = I.right

type t = I.t

Fixed type of containers.

type left = I.left

Fixed type of left elements.

type right = I.right

Fixed type of right elements.

include Bi_mappable_types.Generic with type ('l, 'r) t := t and type 'l left := left and type 'r right := right

include Bi_mappable_types.Basic_generic with type ('l, 'r) t := t with type 'l left := left with type 'r right := right

include Generic_types.Bi_generic with type ('l, 'r) t := t with type 'l left := left with type 'r right := right

val bi_map : t -> left:( left -> left ) -> right:( right -> right ) -> t

bi_map c ~left ~right maps left over every 'l1 left, and right over every 'r1 right, in c.

val map_left : t -> f:( left -> left ) -> t

map_left c ~f maps f over the left type of c only.

val map_right : t -> f:( right -> right ) -> t

map_right c ~f maps f over the right type of c only.