Parameter Map1_left.1-S

include Generic_types.Bi_left

type 'l t

Partially fixed type of containers.

type right

Fixed type of right elements.

include Bi_mappable_types.Generic with type ('l, 'r) t := 'l t and type 'l left := 'l and type 'r right := right

include Bi_mappable_types.Basic_generic with type ('l, 'r) t := 'l t with type 'l left := 'l with type 'r right := right

include Generic_types.Bi_generic with type ('l, 'r) t := 'l t with type 'l left := 'l with type 'r right := right

val bi_map : 'l1 t -> left:( 'l1 -> 'l2 ) -> right:( right -> right ) -> 'l2 t

bi_map c ~left ~right maps left over every 'l1 left, and right over every 'r1 right, in c.

val map_left : 'l1 t -> f:( 'l1 -> 'l2 ) -> 'l2 t

map_left c ~f maps f over the left type of c only.

val map_right : 'l t -> f:( right -> right ) -> 'l t

map_right c ~f maps f over the right type of c only.