Module Bi_mappable.Map1_right

Mapping over the right type of an arity-1 bi-mappable container with a floating right type.

Parameters

S : Bi_mappable_types.S1_right

Signature

type 'r t = 'r S.t

t is the type of the container to map over.

include Mappable_types.Generic with type 'a t := 'a t and type 'a elt := 'a

Generic refers to the container type as 'a t, and the element type as 'a elt; substitute t/elt (arity-0) or 'a t/'a (arity-1) accordingly below.

include Generic_types.Generic

type 'a t

Placeholder for the container type.

type 'a elt

Placeholder for the type of elements inside the container.

val map : 'a t -> f:('a elt -> 'b elt) -> 'b t

map c ~f maps f over every t in c.