Module type Bi_mappable_types.Basic1_right

Basic1_right is the signature of an arity-1 bi-mappable type with a floating right type and fixed left type.

Functions mapped over arity-1 types may change the right element type, but not the left.

include Generic_types.Bi_right

type 'r t

Partially fixed type of containers.

type left

Fixed type of left elements.

include Basic_generic with type ('l, 'r) t := 'r t and type 'l left := left and type 'r right := 'r

include Generic_types.Bi_generic with type ('l, 'r) t := 'r t with type 'l left := left with type 'r right := 'r

val bi_map : 'r1 t -> left:(left -> left) -> right:('r1 -> 'r2) -> 'r2 t

bi_map c ~left ~right maps left over every 'l1 left, and right over every 'r1 right, in c.