Module type Bi_mappable_types.S0

S0 is the full signature of an arity-0 bi-mappable type.

Functions mapped over arity-0 types must preserve both element types.

include Generic_types.Bi0

type t

Fixed type of containers.

type left

Fixed type of left elements.

type right

Fixed type of right elements.

include Generic with type ('l, 'r) t := t and type 'l left := left and type 'r right := right

include Basic_generic with type ('l, 'r) t := t with type 'l left := left with type 'r right := right

include Generic_types.Bi_generic with type ('l, 'r) t := t with type 'l left := left with type 'r right := right

val bi_map : t -> left:(left -> left) -> right:(right -> right) -> t

bi_map c ~left ~right maps left over every 'l1 left, and right over every 'r1 right, in c.

val map_left : t -> f:(left -> left) -> t

map_left c ~f maps f over the left type of c only.

val map_right : t -> f:(right -> right) -> t

map_right c ~f maps f over the right type of c only.