Parameter Chain_Bi2_Traverse1.2-Trav

type 'a t

'a t is the type of the container, parametrised over the element type 'a.

module On (M : Base.Applicative.S) : Traversable_types.S1_on_applicative with type 'a t := 'a t and module M := M

On implements applicative folding and mapping operators for a given applicative M, including arity-1 specific operators.

module On_monad (M : Base.Monad.S) : Traversable_types.S1_on_monad with type 'a t := 'a t and module M := M

On_monad implements applicative folding and mapping operators for a given monad M, including arity-1 specific operators.

module With_errors : Traversable_types.S1_on_monad with type 'a t := 'a t and module M := Base.Or_error

With_errors is shorthand for On_monad (Or_error).

include Traversable_types.Generic with type 'a t := 'a t and type 'a elt := 'a and module On := On and module On_monad := On_monad and module With_errors := With_errors

include Generic_types.Generic with type 'a t := 'a t with type 'a elt := 'a

We can do generic container operations.

include Base.Container.Generic with type 'a t := 'a t and type 'a elt := 'a

We can do non-applicative mapping operations.

include Mappable_types.Generic with type 'a t := 'a t and type 'a elt := 'a

Generic refers to the container type as 'a t, and the element type as 'a elt; substitute t/elt (arity-0) or 'a t/'a (arity-1) accordingly below.

include Generic_types.Generic with type 'a t := 'a t with type 'a elt := 'a

val fold_map : 'a t -> f:( 'acc -> 'a -> 'acc * 'b ) -> init:'acc -> 'acc * 'b t

fold_map c ~f ~init folds f over every t in c, threading through an accumulator with initial value init.

val mapi : f:( Base.int -> 'a -> 'b ) -> 'a t -> 'b t

mapi ~f t maps f across t, passing in an increasing position counter.

include Mappable_types.S1_container with type 'a t := 'a t

include Mappable_types.S1 with type 'a t := 'a t

include Mappable_types.Generic with type 'a t := 'a t and type 'a elt := 'a

Generic refers to the container type as 'a t, and the element type as 'a elt; substitute t/elt (arity-0) or 'a t/'a (arity-1) accordingly below.

include Generic_types.Generic with type 'a t := 'a t with type 'a elt := 'a

val map : 'a t -> f:( 'a -> 'b ) -> 'b t

map c ~f maps f over every t in c.

include Base.Container.S1 with type 'a t := 'a t

val mem : 'a t -> 'a -> equal:( 'a -> 'a -> bool ) -> bool

val length : 'a t -> int

val is_empty : 'a t -> bool

val iter : 'a t -> f:( 'a -> unit ) -> unit

val fold : 'a t -> init:'accum -> f:( 'accum -> 'a -> 'accum ) -> 'accum

val fold_result : 'a t -> init:'accum -> f:( 'accum -> 'a -> ( 'accum, 'e ) Base__.Result.t ) -> ( 'accum, 'e ) Base__.Result.t

val fold_until : 'a t -> init:'accum -> f: ( 'accum -> 'a -> ( 'accum, 'final ) Base__Container_intf.Continue_or_stop.t ) -> finish:( 'accum -> 'final ) -> 'final

val exists : 'a t -> f:( 'a -> bool ) -> bool

val for_all : 'a t -> f:( 'a -> bool ) -> bool

val count : 'a t -> f:( 'a -> bool ) -> int

val sum : (module Base__Container_intf.Summable with type t = 'sum) -> 'a t -> f:( 'a -> 'sum ) -> 'sum

val find : 'a t -> f:( 'a -> bool ) -> 'a option

val find_map : 'a t -> f:( 'a -> 'b option ) -> 'b option

val to_list : 'a t -> 'a list

val to_array : 'a t -> 'a array

val min_elt : 'a t -> compare:( 'a -> 'a -> int ) -> 'a option

val max_elt : 'a t -> compare:( 'a -> 'a -> int ) -> 'a option

include Mappable_types.Extensions1 with type 'a t := 'a t

Extensions1 includes the container extensions from Container_exts, as they work with any arity-1 container.

include Container_exts_types.S1 with type 'a t := 'a t

include Container_exts_types.Generic with type 'a t := 'a t and type 'a elt := 'a

Generic refers to the container type as 'a t, and the element type as 'a elt; substitute t/elt (arity-0) or 'a t/'a (arity-1) accordingly below.

include Generic_types.Generic with type 'a t := 'a t with type 'a elt := 'a

Testing for a specific number of elements

The following functions help in checking whether a container has a particular, commonly-required number of elements (zero or one, one, two, and so on).

val at_most_one : 'a t -> 'a Base.option Base.Or_error.t

at_most_one xs returns Ok None if xs is empty; Ok Some(x) if it contains only x; and an error otherwise.

Examples (using an extended version of List):

(* ok None *)
List.at_most_one []

(* ok (Some 1) *)
List.at_most_one [1]

(* error -- too many *)
List.at_most_one [1; 2]

val one : 'a t -> 'a Base.Or_error.t

one xs returns Ok x if xs contains only x, and an error otherwise.

Examples (using an extended version of List):

(* error -- not enough *)
List.one []

(* ok 1 *)
List.one [1]

(* error -- too many *)
List.one [1; 2]

val two : 'a t -> ('a * 'a) Base.Or_error.t

two xs returns Ok (x, y) if xs is a list containing only x and y in that order, and an error otherwise.

Examples (using an extended version of List):

(* error -- not enough *)
List.two []

(* error -- not enough *)
List.two [1]

(* ok (1, 2) *)
List.two [1; 2]

(* error -- too many *)
List.two [1; 2; 3]

Miscellaneous extensions

val max_measure : measure:( 'a -> Base.int ) -> ?default:Base.int -> 'a t -> Base.int

max_measure ~measure ~default xs measures each item in xs according to measure, and returns the highest measure reported. If xs is empty, return default if given, and 0 otherwise.

Predicate extensions are available on all arity-1 containers, provided that we fix the element type parameter to 'a -> bool.

include Container_exts_types.Generic_predicate with type 'a t := ( 'a -> Base.bool ) t and type 'a item := 'a

val any : 'a -> predicates:( 'a -> Base.bool ) t -> Base.bool

any x ~predicates tests x against predicates until one returns true, in which case it returns true; or all return false, in which case it returns false.

val all : 'a -> predicates:( 'a -> Base.bool ) t -> Base.bool

any x ~predicates tests x against predicates until one returns false, in which case it returns false; or all return true, in which case it returns true.

val none : 'a -> predicates:( 'a -> Base.bool ) t -> Base.bool

none x ~predicates is the same as any x with all predicates in predicates negated. It tests x against predicates until one returns true, in which case it returns false; or all return false, in which case it returns true.

val right_pad : padding:'a -> 'a Base.list t -> 'a Base.list t

right_pad ~padding xs pads every list in xs with padding, ensuring all lists have equal length.

Example:

right_pad ~padding:6
  [ [0; 8; 0; 0] (* [ [ 0; 8; 0; 0; 6 ] *)
  ; [9; 9; 9] (* ; [ 9; 9; 9; 6; 6 ] *)
  ; [8; 8; 1; 9; 9] (* ; [ 8; 8; 1; 9; 9 ] *)
  ; [9; 1; 1; 9] (* ; [ 9; 1; 1; 9; 6 ] *)
  ; [7; 2; 5] (* ; [ 7; 2; 5; 6; 6 ] *)
  ; [3] (* ; [ 3; 6; 6; 6; 6 ] *) ]

(* ] *)