Module Bi_traversable.Chain_Traverse1_Bi1_right

Chain_Traverse1_Bi1_right (RTrav) (Bi) composes an inner arity-1 traversal LTrav on the right of an arity-1 bi-traversal Bi with floating right type.

Parameters

RTrav : Traversable_types.S1

Bi : Bi_traversable_types.Basic1_right

Signature

include Generic_types.Bi_right

type 'r t

Partially fixed type of containers.

type left

Fixed type of left elements.

include Bi_traversable_types.Generic with type ('l, 'r) t := 'r t and type 'l left := left and type 'r right := 'r

include Generic_types.Bi_generic

type ('l, 'r) t

Generic container type.

type 'l left

Generic left type.

type 'r right

Generic right type.

We can do non-monadic bi-mapping operations.

include Bi_mappable_types.Generic with type ('l, 'r) t := ('l, 'r) t and type 'l left := 'l left and type 'r right := 'r right

include Bi_mappable_types.Basic_generic

include Generic_types.Bi_generic

type ('l, 'r) t

Generic container type.

type 'l left

Generic left type.

type 'r right

Generic right type.

val bi_map : ('l1, 'r1) t -> left:('l1 left -> 'l2 left) -> right:('r1 right -> 'r2 right) -> ('l2, 'r2) t

bi_map c ~left ~right maps left over every 'l1 left, and right over every 'r1 right, in c.

val map_left : ('l1, 'r) t -> f:('l1 left -> 'l2 left) -> ('l2, 'r) t

map_left c ~f maps f over the left type of c only.

val map_right : ('l, 'r1) t -> f:('r1 right -> 'r2 right) -> ('l, 'r2) t

map_right c ~f maps f over the right type of c only.

module On : functor (M : Base.Applicative.S) -> Bi_traversable_types.Generic_on_applicative with type ('l, 'r) t := ('l, 'r) t and type 'l left := 'l left and type 'r right := 'r right and module M := M

On implements bi-traversal operators for a given applicative functor M.

module On_monad : functor (M : Base.Monad.S) -> Bi_traversable_types.Generic_on_applicative with type ('l, 'r) t := ('l, 'r) t and type 'l left := 'l left and type 'r right := 'r right and module M := Monad_exts.App(M)

On_monad implements bi-traversal operators for a given monad M.

module With_errors : Bi_traversable_types.Generic_on_applicative with type ('l, 'r) t := ('l, 'r) t and type 'l left := 'l left and type 'r right := 'r right and module M := Base.Or_error

With_errors specialises On_monad to the error_monad.