Module Bi_traversable.Fix2_right

Fix2_right (S) (Left) fixes the right type of S to Right, making it an S1_left.

Parameters

module I : Bi_traversable_types.Basic2

module Right : Base.T

Signature

include Generic_types.Bi_left with type 'l t = ( 'l, Right.t ) I.t with type right = Right.t

type 'l t = ( 'l, Right.t ) I.t

Partially fixed type of containers.

type right = Right.t

Fixed type of right elements.

include Bi_traversable_types.Generic with type ('l, 'r) t := 'l t and type 'l left := 'l and type 'r right := right

include Generic_types.Bi_generic with type ('l, 'r) t := 'l t with type 'l left := 'l with type 'r right := right

We can do non-monadic bi-mapping operations.

include Bi_mappable_types.Generic with type ('l, 'r) t := 'l t and type 'l left := 'l and type 'r right := right

include Bi_mappable_types.Basic_generic with type ('l, 'r) t := 'l t with type 'l left := 'l with type 'r right := right

include Generic_types.Bi_generic with type ('l, 'r) t := 'l t with type 'l left := 'l with type 'r right := right

val bi_map : 'l1 t -> left:( 'l1 -> 'l2 ) -> right:( right -> right ) -> 'l2 t

bi_map c ~left ~right maps left over every 'l1 left, and right over every 'r1 right, in c.

val map_left : 'l1 t -> f:( 'l1 -> 'l2 ) -> 'l2 t

map_left c ~f maps f over the left type of c only.

val map_right : 'l t -> f:( right -> right ) -> 'l t

map_right c ~f maps f over the right type of c only.

module On (M : Base.Applicative.S) : Bi_traversable_types.Generic_on_applicative with type ('l, 'r) t := 'l t and type 'l left := 'l and type 'r right := right and module M := M

On implements bi-traversal operators for a given applicative functor M.

module On_monad (M : Base.Monad.S) : Bi_traversable_types.Generic_on_applicative with type ('l, 'r) t := 'l t and type 'l left := 'l and type 'r right := right and module M := Monad_exts.App(M)

On_monad implements bi-traversal operators for a given monad M.

module With_errors : Bi_traversable_types.Generic_on_applicative with type ('l, 'r) t := 'l t and type 'l left := 'l and type 'r right := right and module M := Base.Or_error

With_errors specialises On_monad to the error_monad.