Module Bi_traversable.Traverse0_left

Traversing over the left type of an arity-0 bi-traversable container, which must have equality as witnessed by an Equal.S module.

Parameters

module L : Base.Equal.S

module S : Bi_traversable_types.S0 with type left := L.t

Signature

module Elt = L

Elements must have equality. While this is an extra restriction on top of the Base equivalent, it is required by Make0, and helps us define chaining operations.

We export Elt.t as elt for compatibility with Core-style containers.

include Generic_types.S0 with type elt = Elt.t with type t = S.t

type t = S.t

The container type.

type elt = Elt.t

The element type.

include Traversable_types.Generic with type ('a, 'phantom) t := t and type 'a elt := Elt.t

include Generic_types.Generic with type ('a, 'phantom) t := t with type 'a elt := Elt.t

module On (M : Base.Applicative.S) : Traversable_types.Generic_on_applicative with type ('a, 'phantom) t := t and type 'a elt := Elt.t and module M := M

On implements traversal operators for a given applicative M.

module On_monad (M : Base.Monad.S) : Traversable_types.Generic_on_monad with type ('a, 'phantom) t := t and type 'a elt := Elt.t and module M := M

On_monad implements traversal operators for a given monad M. Compared to On(Monad_exts.App(M)), this adds various derived operators available only for monads.

We can do generic container operations.

include Base.Container.Generic with type ('a, 'phantom) t := t and type 'a elt := Elt.t

We can do non-applicative mapping operations.

include Mappable_types.Generic with type ('a, 'phantom) t := t and type 'a elt := Elt.t

Generic refers to the container type as 'a t, and the element type as 'a elt; substitute t/elt (arity-0) or 'a t/'a (arity-1) accordingly below.

include Generic_types.Generic with type ('a, 'phantom) t := t with type 'a elt := Elt.t

val fold_map : t -> f:('acc -> Elt.t -> 'acc * Elt.t) -> init:'acc -> 'acc * t

fold_map c ~f ~init folds f over every t in c, threading through an accumulator with initial value init.

val mapi : f:(Base.int -> Elt.t -> Elt.t) -> t -> t

mapi ~f t maps f across t, passing in an increasing position counter.

module With_errors : Traversable_types.Generic_on_monad with type ('a, 'phantom) t := t and type 'a elt := Elt.t and module M := Base.Or_error

With_errors specialises On_applicative to the error applicative.

include Mappable_types.S0_container with type t := t and type elt := Elt.t

include Mappable_types.S0 with type t := t with type elt := Elt.t

include Generic_types.S0 with type t := t with type elt := Elt.t

include Mappable_types.Generic with type ('a, 'phantom) t := t and type 'a elt := Elt.t

Generic refers to the container type as 'a t, and the element type as 'a elt; substitute t/elt (arity-0) or 'a t/'a (arity-1) accordingly below.

include Generic_types.Generic with type ('a, 'phantom) t := t with type 'a elt := Elt.t

val map : t -> f:(Elt.t -> Elt.t) -> t

map c ~f maps f over every t in c.

include Base.Container.S0 with type t := t and type elt := Elt.t

val mem : t -> Elt.t -> bool

val length : t -> int

val is_empty : t -> bool

val iter : t -> f:(Elt.t -> unit) -> unit

val fold : t -> init:'acc -> f:('acc -> Elt.t -> 'acc) -> 'acc

val fold_result : t -> init:'acc -> f:('acc -> Elt.t -> ('acc, 'e) Base__.Result.t) -> ('acc, 'e) Base__.Result.t

val fold_until : t -> init:'acc -> f:('acc -> Elt.t -> ('acc, 'final) Base__Container_intf.Continue_or_stop.t) -> finish:('acc -> 'final) -> 'final

val exists : t -> f:(Elt.t -> bool) -> bool

val for_all : t -> f:(Elt.t -> bool) -> bool

val count : t -> f:(Elt.t -> bool) -> int

val sum : (module Base__Container_intf.Summable with type t = 'sum) -> t -> f:(Elt.t -> 'sum) -> 'sum

val find : t -> f:(Elt.t -> bool) -> Elt.t option

val find_map : t -> f:(Elt.t -> 'a option) -> 'a option

val to_list : t -> Elt.t list

val to_array : t -> Elt.t array

val min_elt : t -> compare:(Elt.t -> Elt.t -> int) -> Elt.t option

val max_elt : t -> compare:(Elt.t -> Elt.t -> int) -> Elt.t option