Module Traverse1_right.On

On implements applicative folding and mapping operators for a given applicative M, including arity-1 specific operators.

Parameters

module M : Base.Applicative.S

Signature

include Traversable_types.Generic_on_applicative with type ('a, 'phantom) t := 'a t and type 'a elt := 'a with module M := M

include Traversable_types.Basic_generic_on_applicative with type ('a, 'phantom) t := 'a t with type 'a elt := 'a with module M := M

Generic refers to the container type as 'a t, and the element type as 'a elt; substitute t/elt (arity-0) or 'a t/'a (arity-1) accordingly below.

include Generic_types.Generic with type ('a, 'phantom) t := 'a t with type 'a elt := 'a

val map_m : 'a t -> f:('a -> 'b M.t) -> 'b t M.t

map_m c ~f maps f over every t in c, threading through an applicative functor.

Example:

(* Travesty_base_exts.List adds applicative traversals to a list;
   With_errors (in S1_container) implements them on the On_error
   applicative functor. *)

let f x =
  Or_error.(if 0 < x then error_string "negative!" else ok x)
in
List.With_errors.map_m integers ~f

val iter_m : 'a t -> f:('a -> Base.unit M.t) -> Base.unit M.t

iter_m x ~f iterates the computation f over x, returning the final applicative effect.

val sequence_m : 'a M.t t -> 'a t M.t

sequence_m x lifts a container of applicatives x to an applicative containing a container, by sequencing the applicative effects from left to right.