Module type Bi_traversable_types.Basic1_left

Basic1_left is the basic signature of an arity-1 bi-traversable type with a floating left type and fixed right type.

Functions traversing over arity-1 types may change the left element type, but not the right.

include Generic_types.Bi_left

type 'l t

Partially fixed type of containers.

type right

Fixed type of right elements.

module On (M : Base.Applicative.S) : Basic1_left_on_applicative with type 'r t := 'r t and type right := right and module M := M

On implements monadic bi-traversal for a given applicative functor.