Module type Bi_traversable_types.Basic1_left_on_applicative

Basic1_left_on_applicative is the inner signature of an applicative bi-traversal over arity-1 types with a floating left type and fixed right type.

include Generic_types.Bi_left

type 'l t

Partially fixed type of containers.

type right

Fixed type of right elements.

include Basic_generic_on_applicative with type ('l, 'r) t := 'l t and type 'l left := 'l and type 'r right := right

include Generic_types.Bi_generic with type ('l, 'r) t := 'l t with type 'l left := 'l with type 'r right := right

module M : Base.Applicative.S

M is the applicative functor over which we're bi-traversing.

val bi_map_m : 'l1 t -> left:( 'l1 -> 'l2 M.t ) -> right:( right -> right M.t ) -> 'l2 t M.t

bi_map_m c ~left ~right traverses with left over every 'l1 left, and right over every 'r1 right, in c.