Module type Bi_traversable_types.Basic2_on_applicative

Basic2_on_applicative is the inner signature of a monadic bi-traversal over arity-2 types with a floating right type and fixed left type.

include Base.T2

type ('a, 'b) t

include Basic_generic_on_applicative with type ('l, 'r) t := ('l, 'r) t and type 'l left := 'l and type 'r right := 'r

include Generic_types.Bi_generic with type ('l, 'r) t := ('l, 'r) t with type 'l left := 'l with type 'r right := 'r

module M : Base.Applicative.S

M is the applicative functor over which we're bi-traversing.

val bi_map_m : ('l1, 'r1) t -> left:('l1 -> 'l2 M.t) -> right:('r1 -> 'r2 M.t) -> ('l2, 'r2) t M.t

bi_map_m c ~left ~right traverses with left over every 'l1 left, and right over every 'r1 right, in c.