Module type Bi_traversable_types.Generic_on_applicative

Generic_on_applicative extends Generic to contain various derived operators; we use it to derive the signatures of the various On modules.

For arity-0 types, ('l, 'r) t becomes t, 'l left becomes left, and 'r right becomes right.

For arity-1 types with a fixed right type, ('l, 'r) t becomes 'l t, 'l left becomes 'l, and 'r right becomes right.

For arity-1 types with a fixed left type, ('l, 'r) t becomes 'l t, 'l left becomes left, and 'r right becomes 'r.

For arity-2 types, ('l, 'r) t becomes ('l, 'r) t, 'l left becomes 'l, and 'r right becomes 'r.

include Basic_generic_on_applicative

include Generic_types.Bi_generic

type ('l, 'r) t

Generic container type.

type 'l left

Generic left type.

type 'r right

Generic right type.

module M : Base.Applicative.S

M is the applicative functor over which we're bi-traversing.

val bi_map_m : ( 'l1, 'r1 ) t -> left:( 'l1 left -> 'l2 left M.t ) -> right:( 'r1 right -> 'r2 right M.t ) -> ( 'l2, 'r2 ) t M.t

bi_map_m c ~left ~right traverses with left over every 'l1 left, and right over every 'r1 right, in c.

val map_left_m : ( 'l1, 'r ) t -> f:( 'l1 left -> 'l2 left M.t ) -> ( 'l2, 'r ) t M.t

map_left_m c ~f monadically traverses over the left type of c only, using f.

val map_right_m : ( 'l, 'r1 ) t -> f:( 'r1 right -> 'r2 right M.t ) -> ( 'l, 'r2 ) t M.t

map_right_m c ~f monadically traverses f over the right type of c only, using f.