Module S2.On_monad

On_monad implements bi-traversal operators for a given monad M.

Currently, this is no different to On(Monad_exts.App(M)); this may change in the future.

Parameters

M : Base.Monad.S

Signature

include Basic_generic_on_applicative

include Generic_types.Bi_generic

type ('l, 'r) t

Generic container type.

type 'l left

Generic left type.

type 'r right

Generic right type.

module M : Base.Applicative.S

M is the applicative functor over which we're bi-traversing.

val bi_map_m : ('l1, 'r1) t -> left:('l1 left -> 'l2 left M.t) -> right:('r1 right -> 'r2 right M.t) -> ('l2, 'r2) t M.t

bi_map_m c ~left ~right traverses with left over every 'l1 left, and right over every 'r1 right, in c.

val map_left_m : ('l1, 'r) t -> f:('l1 left -> 'l2 left M.t) -> ('l2, 'r) t M.t

map_left_m c ~f monadically traverses over the left type of c only, using f.

val map_right_m : ('l, 'r1) t -> f:('r1 right -> 'r2 right M.t) -> ('l, 'r2) t M.t

map_right_m c ~f monadically traverses f over the right type of c only, using f.