Module Container_exts.Extend0_predicate

Extend0_predicate creates extensions for a Container.S0 over predicates.

Parameters

module P : Base.T

module C : Base.Container.S0 with type elt = P.t -> Base.bool

Signature

include Container_exts_types.S0 with type t := C.t and type elt := P.t -> Base.bool

include Generic_types.S0 with type t := C.t with type elt := P.t -> Base.bool

include Container_exts_types.Generic with type 'a t := C.t and type 'a elt := P.t -> Base.bool

Generic refers to the container type as 'a t, and the element type as 'a elt; substitute t/elt (arity-0) or 'a t/'a (arity-1) accordingly below.

include Generic_types.Generic with type 'a t := C.t with type 'a elt := P.t -> Base.bool

Testing for a specific number of elements

The following functions help in checking whether a container has a particular, commonly-required number of elements (zero or one, one, two, and so on).

val at_most_one : C.t -> ( P.t -> Base.bool ) Base.option Base.Or_error.t

at_most_one xs returns Ok None if xs is empty; Ok Some(x) if it contains only x; and an error otherwise.

Examples (using an extended version of List):

(* ok None *)
List.at_most_one []

(* ok (Some 1) *)
List.at_most_one [1]

(* error -- too many *)
List.at_most_one [1; 2]

val one : C.t -> ( P.t -> Base.bool ) Base.Or_error.t

one xs returns Ok x if xs contains only x, and an error otherwise.

Examples (using an extended version of List):

(* error -- not enough *)
List.one []

(* ok 1 *)
List.one [1]

(* error -- too many *)
List.one [1; 2]

val two : C.t -> (( P.t -> Base.bool ) * ( P.t -> Base.bool )) Base.Or_error.t

two xs returns Ok (x, y) if xs is a list containing only x and y in that order, and an error otherwise.

Examples (using an extended version of List):

(* error -- not enough *)
List.two []

(* error -- not enough *)
List.two [1]

(* ok (1, 2) *)
List.two [1; 2]

(* error -- too many *)
List.two [1; 2; 3]

Miscellaneous extensions

val max_measure : measure:( ( P.t -> Base.bool ) -> Base.int ) -> ?default:Base.int -> C.t -> Base.int

max_measure ~measure ~default xs measures each item in xs according to measure, and returns the highest measure reported. If xs is empty, return default if given, and 0 otherwise.

include Container_exts_types.Generic_predicate with type 'a t := C.t and type 'a item := P.t

val any : P.t -> predicates:C.t -> Base.bool

any x ~predicates tests x against predicates until one returns true, in which case it returns true; or all return false, in which case it returns false.

val all : P.t -> predicates:C.t -> Base.bool

any x ~predicates tests x against predicates until one returns false, in which case it returns false; or all return true, in which case it returns true.

val none : P.t -> predicates:C.t -> Base.bool

none x ~predicates is the same as any x with all predicates in predicates negated. It tests x against predicates until one returns true, in which case it returns false; or all return false, in which case it returns true.