Module Travesty.Container_exts_types

Extension signatures for containers.

This module contains various signatures that enumerate extensions that apply to Base containers of both arity 0 and arity 1.

Generic extensions

As is often the case in Travesty, we define an arity-generic signature first, then specialise it for arity-0 and arity-1 containers.

module type Generic = sig ... end

Containers of predicates

The following functions concern containers of predicates (functions of type 'a -> bool).

module type Generic_predicate = sig ... end

Arity-0 container extensions

These extensions target arity-0 containers (implementations of Container.S0).

module type S0 = sig ... end

Extensions for a Container.S0.

module type S0_predicate = sig ... end

Extensions for a Container.S0 whose elements are predicates.

Arity-1 container extensions

These extensions target arity-1 containers (implementations of Container.S1).

module type S1 = sig ... end

Extensions for a Container.S1.