Module type Container_exts_types.Generic_predicate

type 'a t

The generic type of predicate containers.

type 'a item

The generic type of predicate target elements.

val any : 'a item -> predicates:'a t -> Base.bool

any x ~predicates tests x against predicates until one returns true, in which case it returns true; or all return false, in which case it returns false.

val all : 'a item -> predicates:'a t -> Base.bool

any x ~predicates tests x against predicates until one returns false, in which case it returns false; or all return true, in which case it returns true.

val none : 'a item -> predicates:'a t -> Base.bool

none x ~predicates is the same as any x with all predicates in predicates negated. It tests x against predicates until one returns true, in which case it returns false; or all return false, in which case it returns true.