Module type Container_exts_types.S0

Extensions for a Container.S0.

This signature contains the generic extensions outlined in Generic.

include Generic_types.S0

type t

The container type.

type elt

The element type.

include Generic with type 'a t := t and type 'a elt := elt

Generic refers to the container type as 'a t, and the element type as 'a elt; substitute t/elt (arity-0) or 'a t/'a (arity-1) accordingly below.

include Generic_types.Generic with type 'a t := t with type 'a elt := elt

Testing for a specific number of elements

The following functions help in checking whether a container has a particular, commonly-required number of elements (zero or one, one, two, and so on).

val at_most_one : t -> elt Base.option Base.Or_error.t

at_most_one xs returns Ok None if xs is empty; Ok Some(x) if it contains only x; and an error otherwise.

Examples (using an extended version of List):

(* ok None *)
List.at_most_one []

(* ok (Some 1) *)
List.at_most_one [1]

(* error -- too many *)
List.at_most_one [1; 2]

val one : t -> elt Base.Or_error.t

one xs returns Ok x if xs contains only x, and an error otherwise.

Examples (using an extended version of List):

(* error -- not enough *)
List.one []

(* ok 1 *)
List.one [1]

(* error -- too many *)
List.one [1; 2]

val two : t -> (elt * elt) Base.Or_error.t

two xs returns Ok (x, y) if xs is a list containing only x and y in that order, and an error otherwise.

Examples (using an extended version of List):

(* error -- not enough *)
List.two []

(* error -- not enough *)
List.two [1]

(* ok (1, 2) *)
List.two [1; 2]

(* error -- too many *)
List.two [1; 2; 3]

Miscellaneous extensions

val max_measure : measure:( elt -> Base.int ) -> ?default:Base.int -> t -> Base.int

max_measure ~measure ~default xs measures each item in xs according to measure, and returns the highest measure reported. If xs is empty, return default if given, and 0 otherwise.