Module Filter_mappable.Make1

Make0 makes an S1 from a Basic1.

Parameters

module F : Filter_mappable_types.Basic1

Signature

include Filter_mappable_types.Generic with type ('a, 'phantom) t := 'a F.t and type 'a elt := 'a

Generic strictly extends Generic_basic.

include Filter_mappable_types.Generic_basic with type ('a, 'phantom) t := 'a F.t with type 'a elt := 'a

Generic_basic refers to the container type as 'a t, and the element type as 'a elt; substitute t/elt (arity-0) or 'a t/'a (arity-1) accordingly below.

include Generic_types.Generic with type ('a, 'phantom) t := 'a F.t with type 'a elt := 'a

val filter_map : 'a F.t -> f:('a -> 'b Base.option) -> 'b F.t

filter_map c ~f maps f over every t in c, discarding any items for which f returns None.

val filter : 'a F.t -> f:('a -> Base.bool) -> 'a F.t

filter c ~f checks f over every t in c, discarding any items for which f returns false.

val exclude : 'a F.t -> f:('a -> Base.bool) -> 'a F.t

exclude c ~f checks f over every t in c, discarding any items for which f returns true.

We can also derive Mappable interfaces from filter-mappable ones, but leave that to a separate functor.