Module type Filter_mappable_types.Generic_basic

Generic input shape of filter-mappable structures on either an arity-0 or arity-1 type.

For arity-0 types, use Basic0: 'a t becomes t, and 'a elt becomes elt;

For arity-1 types, use Basic1: 'a t becomes 'a t, and 'a elt becomes 'a.

Generic_basic refers to the container type as 'a t, and the element type as 'a elt; substitute t/elt (arity-0) or 'a t/'a (arity-1) accordingly below.

include Generic_types.Generic

type 'a t

Placeholder for the container type.

type 'a elt

Placeholder for the type of elements inside the container.

val filter_map : 'a t -> f:( 'a elt -> 'b elt Base.option ) -> 'b t

filter_map c ~f maps f over every t in c, discarding any items for which f returns None.