Module type Filter_mappable_types.S1

Full signature of an arity-1 mappable type.

Functions mapped over arity-1 types may change the element type.

type 'a t

The type of the container to map over.

include Generic with type 'a t := 'a t and type 'a elt := 'a

Generic strictly extends Generic_basic.

include Generic_basic with type 'a t := 'a t with type 'a elt := 'a

Generic_basic refers to the container type as 'a t, and the element type as 'a elt; substitute t/elt (arity-0) or 'a t/'a (arity-1) accordingly below.

include Generic_types.Generic with type 'a t := 'a t with type 'a elt := 'a

val filter_map : 'a t -> f:( 'a -> 'b Base.option ) -> 'b t

filter_map c ~f maps f over every t in c, discarding any items for which f returns None.

val filter : 'a t -> f:( 'a -> Base.bool ) -> 'a t

filter c ~f checks f over every t in c, discarding any items for which f returns false.

val exclude : 'a t -> f:( 'a -> Base.bool ) -> 'a t

exclude c ~f checks f over every t in c, discarding any items for which f returns true.

We can also derive Mappable interfaces from filter-mappable ones, but leave that to a separate functor.