Module Mappable.Extend1

Extend1 implements Extensions1 for an arity-1 mappable container.

Parameters

module S : Mappable_types.S1_container

Signature

Extensions1 includes the container extensions from Container_exts, as they work with any arity-1 container.

include Container_exts_types.S1 with type 'a t := 'a S.t

include Container_exts_types.Generic with type ('a, 'phantom) t := 'a S.t and type 'a elt := 'a

Generic refers to the container type as 'a t, and the element type as 'a elt; substitute t/elt (arity-0) or 'a t/'a (arity-1) accordingly below.

include Generic_types.Generic with type ('a, 'phantom) t := 'a S.t with type 'a elt := 'a

Testing for a specific number of elements

The following functions help in checking whether a container has a particular, commonly-required number of elements (zero or one, one, two, and so on).

val at_most_one : 'a S.t -> 'a Base.option Base.Or_error.t

at_most_one xs returns Ok None if xs is empty; Ok Some(x) if it contains only x; and an error otherwise.

Examples (using an extended version of List):

(* ok None *)
List.at_most_one []

(* ok (Some 1) *)
List.at_most_one [1]

(* error -- too many *)
List.at_most_one [1; 2]

val one : 'a S.t -> 'a Base.Or_error.t

one xs returns Ok x if xs contains only x, and an error otherwise.

Examples (using an extended version of List):

(* error -- not enough *)
List.one []

(* ok 1 *)
List.one [1]

(* error -- too many *)
List.one [1; 2]

val two : 'a S.t -> ('a * 'a) Base.Or_error.t

two xs returns Ok (x, y) if xs is a list containing only x and y in that order, and an error otherwise.

Examples (using an extended version of List):

(* error -- not enough *)
List.two []

(* error -- not enough *)
List.two [1]

(* ok (1, 2) *)
List.two [1; 2]

(* error -- too many *)
List.two [1; 2; 3]

Miscellaneous extensions

val max_measure : measure:('a -> Base.int) -> ?default:Base.int -> 'a S.t -> Base.int

max_measure ~measure ~default xs measures each item in xs according to measure, and returns the highest measure reported. If xs is empty, return default if given, and 0 otherwise.

Predicate extensions are available on all arity-1 containers, provided that we fix the element type parameter to 'a -> bool.

include Container_exts_types.Generic_predicate with type ('a, 'phantom) t := ('a -> Base.bool) S.t and type 'a item := 'a

val any : 'a -> predicates:('a -> Base.bool) S.t -> Base.bool

any x ~predicates tests x against predicates until one returns true, in which case it returns true; or all return false, in which case it returns false.

val all : 'a -> predicates:('a -> Base.bool) S.t -> Base.bool

any x ~predicates tests x against predicates until one returns false, in which case it returns false; or all return true, in which case it returns true.

val none : 'a -> predicates:('a -> Base.bool) S.t -> Base.bool

none x ~predicates is the same as any x with all predicates in predicates negated. It tests x against predicates until one returns true, in which case it returns false; or all return false, in which case it returns true.

val right_pad : padding:'a -> 'a Base.list S.t -> 'a Base.list S.t

right_pad ~padding xs pads every list in xs with padding, ensuring all lists have equal length.

Example:

right_pad ~padding:6
  [ [0; 8; 0; 0] (* [ [ 0; 8; 0; 0; 6 ] *)
  ; [9; 9; 9] (* ; [ 9; 9; 9; 6; 6 ] *)
  ; [8; 8; 1; 9; 9] (* ; [ 8; 8; 1; 9; 9 ] *)
  ; [9; 1; 1; 9] (* ; [ 9; 1; 1; 9; 6 ] *)
  ; [7; 2; 5] (* ; [ 7; 2; 5; 6; 6 ] *)
  ; [3] (* ; [ 3; 6; 6; 6; 6 ] *) ]

(* ] *)