Module Travesty.Monad_transform

Monad transformer signatures.

These signatures capture the monad transformer pattern in a generic way.

module type S_fixed = sig ... end

S_fixed is the signature of monad transformers fixed to a particular inner monad.

module type S = sig ... end

S is the signature of monad transformers.