Module S.On_monad

On_monad is a functor that transforms its argument monad.

Parameters

module M : Base.Monad.S

Signature

val lift : 'a M.t -> 'a M.t t

lift x lifts x from the inner monad to the outer one.