Module Travesty.State_transform_types

module type Generic_types = sig ... end

Generic_types contains generic versions of the types used in Generic_builders and Generic_runners.

module type Generic_builders = sig ... end

Generic_builders contains generic versions of the 'builder' functions common to all state monad signatures.

module type Generic_runners = sig ... end

Generic_runners contains generic versions of the 'runner' functions common to all state monad signatures.

module type Fix = sig ... end

Fix contains the signature for fixpoint builders.

module type Generic = sig ... end

Generic contains the signature bits common to all state transformers.

module type S = sig ... end

S is the signature of state monad transformers with a fixed state type.

module type Basic = sig ... end

Basic is the signature that must be implemented by state systems being lifted into S_transform instances.

module type S2 = sig ... end

S2 is the signature of state transformers parametrised over both value and state types.