Module type State_transform_types.Generic_types

Generic_types contains generic versions of the types used in Generic_builders and Generic_runners.

type ('a, 's) t

t is the type of the state monad.

type 'a final

final is the type of returned results. In transformers, this becomes 'a Inner.t; otherwise, it becomes just 'a.

type 's state

state is the type used to represent the state outside of its monad. In S, 's state becomes x for some type x; in S2, 's state becomes 's.