Parameter Chain0.2-Inner

module Elt : Base.Equal.S

Elements must have equality. While this is an extra restriction on top of the Core equivalent, it is required by Make0, and helps us define chaining operations.

We export Elt.t as elt for compatibility with Core-style containers.

include Generic_types.S0 with type elt = Elt.t

type t

The container type.

type elt = Elt.t

The element type.

include Traversable_types.Generic with type 'a t := t and type 'a elt := Elt.t

include Generic_types.Generic

type 'a t

Placeholder for the container type.

type 'a elt

Placeholder for the type of elements inside the container.

module On_monad : functor (M : Base.Monad.S) -> Traversable_types.Generic_on_monad with type 'a t := 'a t and type 'a elt := 'a elt and module M := M

On_monad implements monadic traversal operators for a given monad M.

We can do generic container operations.

include Base.Container.Generic with type 'a Inner.t := 'a t and type 'a Inner.elt := 'a elt

val length : 'a t -> int

val is_empty : 'a t -> bool

val iter : 'a t -> f:('a elt -> unit) -> unit

val fold : 'a t -> init:'accum -> f:('accum -> 'a elt -> 'accum) -> 'accum

val fold_result : 'a t -> init:'accum -> f:('accum -> 'a elt -> ('accum, 'e) Base__.Result.t) -> ('accum, 'e) Base__.Result.t

val fold_until : 'a t -> init:'accum -> f:('accum -> 'a elt -> ('accum, 'final) Base__.Container_intf.Continue_or_stop.t) -> finish:('accum -> 'final) -> 'final

val exists : 'a t -> f:('a elt -> bool) -> bool

val for_all : 'a t -> f:('a elt -> bool) -> bool

val count : 'a t -> f:('a elt -> bool) -> int

val sum : (module Base__.Container_intf.Summable with type t = 'sum) -> 'a t -> f:('a elt -> 'sum) -> 'sum

val find : 'a t -> f:('a elt -> bool) -> 'a elt option

val find_map : 'a t -> f:('a elt -> 'b option) -> 'b option

val to_list : 'a t -> 'a elt list

val to_array : 'a t -> 'a elt array

val min_elt : 'a t -> compare:('a elt -> 'a elt -> int) -> 'a elt option

val max_elt : 'a t -> compare:('a elt -> 'a elt -> int) -> 'a elt option

We can do non-monadic mapping operations.

include Mappable_types.Generic with type 'a t := 'a t and type 'a elt := 'a elt

Generic refers to the container type as 'a t, and the element type as 'a elt; substitute t/elt (arity-0) or 'a t/'a (arity-1) accordingly below.

include Generic_types.Generic

type 'a t

Placeholder for the container type.

type 'a elt

Placeholder for the type of elements inside the container.

val map : 'a t -> f:('a elt -> 'b elt) -> 'b t

map c ~f maps f over every t in c.

val fold_map : 'a t -> f:('acc -> 'a elt -> 'acc * 'b elt) -> init:'acc -> 'acc * 'b t

fold_map c ~f ~init folds f over every t in c, threading through an accumulator with initial value init.

val mapi : f:(Base.int -> 'a elt -> 'b elt) -> 'a t -> 'b t

mapi ~f t maps f across t, passing in an increasing position counter.

module With_errors : Traversable_types.Generic_on_monad with type 'a t := 'a t and type 'a elt := 'a elt and module M := Base.Or_error

With_errors specialises On_monad to the error monad.

include Mappable_types.S0_container with type t := t and type elt := Elt.t

include Mappable_types.S0

include Generic_types.S0

type t

The container type.

type elt

The element type.

include Mappable_types.Generic with type 'a t := t and type 'a elt := elt

Generic refers to the container type as 'a t, and the element type as 'a elt; substitute t/elt (arity-0) or 'a t/'a (arity-1) accordingly below.

include Generic_types.Generic

type 'a t

Placeholder for the container type.

type 'a elt

Placeholder for the type of elements inside the container.

val map : 'a t -> f:('a elt -> 'b elt) -> 'b t

map c ~f maps f over every t in c.

include Base.Container.S0 with type Inner.t := t and type Inner.elt := elt

val mem : t -> elt -> bool

val length : t -> int

val is_empty : t -> bool

val iter : t -> f:(elt -> unit) -> unit

val fold : t -> init:'accum -> f:('accum -> elt -> 'accum) -> 'accum

val fold_result : t -> init:'accum -> f:('accum -> elt -> ('accum, 'e) Base__.Result.t) -> ('accum, 'e) Base__.Result.t

val fold_until : t -> init:'accum -> f:('accum -> elt -> ('accum, 'final) Base__.Container_intf.Continue_or_stop.t) -> finish:('accum -> 'final) -> 'final

val exists : t -> f:(elt -> bool) -> bool

val for_all : t -> f:(elt -> bool) -> bool

val count : t -> f:(elt -> bool) -> int

val sum : (module Base__.Container_intf.Summable with type t = 'sum) -> t -> f:(elt -> 'sum) -> 'sum

val find : t -> f:(elt -> bool) -> elt option

val find_map : t -> f:(elt -> 'a option) -> 'a option

val to_list : t -> elt list

val to_array : t -> elt array

val min_elt : t -> compare:(elt -> elt -> int) -> elt option

val max_elt : t -> compare:(elt -> elt -> int) -> elt option