Module Fix_elt.On_monad

On_monad implements traversal operators for a given monad M. Compared to On(Monad_exts.App(M)), this adds various derived operators available only for monads.

Parameters

module M : Base.Monad.S

Signature

All applicative operators are available through lowering the monad to an applicative functor.

include Traversable_types.Generic_on_applicative with module M := Monad_exts.App(M) with type ('a, 'phantom) t := t with type 'a elt := Elt.t

include Traversable_types.Basic_generic_on_applicative with module M := Monad_exts.App(M) with type ('a, 'phantom) t := t with type 'a elt := Elt.t

Generic refers to the container type as 'a t, and the element type as 'a elt; substitute t/elt (arity-0) or 'a t/'a (arity-1) accordingly below.

include Generic_types.Generic with type ('a, 'phantom) t := t with type 'a elt := Elt.t

val map_m : t -> f:(Elt.t -> Elt.t Travesty.Monad_exts.App(M).t) -> t Travesty.Monad_exts.App(M).t

map_m c ~f maps f over every t in c, threading through an applicative functor.

Example:

(* Travesty_base_exts.List adds applicative traversals to a list;
   With_errors (in S1_container) implements them on the On_error
   applicative functor. *)

let f x =
  Or_error.(if 0 < x then error_string "negative!" else ok x)
in
List.With_errors.map_m integers ~f

val fold_map_m : t -> f:('acc -> Elt.t -> ('acc * Elt.t) M.t) -> init:'acc -> ('acc * t) M.t

fold_map_m c ~f ~init folds f applicatively over every t in c, threading through an accumulator with initial value init.

val fold_m : t -> init:'acc -> f:('acc -> Elt.t -> 'acc M.t) -> 'acc M.t

fold_m x ~init ~f folds the applicative computation f over x, starting with initial value init, and returning the final value inside the applicative effect.

val iter_m : t -> f:(Elt.t -> Base.unit M.t) -> Base.unit M.t

iter_m x ~f iterates the computation f over x, returning the final applicative effect.

val mapi_m : f:(Base.int -> Elt.t -> Elt.t M.t) -> t -> t M.t

mapi_m ~f x behaves as map_m, but also supplies f with the index of the element. This index should match the actual position of the element in the container x.