Parameter Make0.I

type t

The container type.

module Elt : Base.Equal.S

Elt contains the element type, which must have equality.

module On (M : Base.Applicative.S) : Traversable_types.Basic0_on_applicative with type t := t and type elt := Elt.t and module M := M

On implements applicative traversal for a given applicative M.