Module 1-I.On_monad

On_monad implements monadic traversal for a given monad.

Parameters

M : Base.Monad.S

Signature

type 'a t

The type of the container to map over.

include Traversable_types.Basic_generic_on_monad with type 'a t := 'a t and type 'a elt := 'a

Generic refers to the container type as 'a t, and the element type as 'a elt; substitute t/elt (arity-0) or 'a t/'a (arity-1) accordingly below.

include Generic_types.Generic

type 'a t

Placeholder for the container type.

type 'a elt

Placeholder for the type of elements inside the container.

module M : Base.Monad.S

M is the monad over which we're fold-mapping.

val map_m : 'a t -> f:('a elt -> 'b elt M.t) -> 'b t M.t

map_m c ~f maps f over every t in c, threading through monadic state.

Example:

(* Travesty_base_exts.List adds monadic traversals to a list;
   With_errors (in S1_container) implements them on the On_error
   monad. *)

let f x =
  Or_error.(if 0 < x then error_string "negative!" else ok x)
in
List.With_errors.map_m integers ~f