Module type Traversable_types.Basic0

Basic0 is the minimal signature that traversable containers of arity 0 must implement to be extensible into S0.

type t

The container type.

module Elt : Base.Equal.S

Elt contains the element type, which must have equality.

module On : functor (M : Base.Applicative.S) -> Basic0_on_applicative with type t := t and type elt := Elt.t and module M := M

On implements applicative traversal for a given applicative M.