Module type Traversable_types.Basic0_container

Basic0_container combines Basic0 and the Base container signature, and is used for extending existing containers into S0_containers.

include Basic0

type t

The container type.

module Elt : Base.Equal.S

Elt contains the element type, which must have equality.

module On (M : Base.Applicative.S) : Basic0_on_applicative with type t := t and type elt := Elt.t and module M := M

On implements applicative traversal for a given applicative M.

include Base.Container.S0 with type t := t and type elt := Elt.t

val mem : t -> Elt.t -> bool

val length : t -> int

val is_empty : t -> bool

val iter : t -> f:(Elt.t -> unit) -> unit

val fold : t -> init:'acc -> f:('acc -> Elt.t -> 'acc) -> 'acc

val fold_result : t -> init:'acc -> f:('acc -> Elt.t -> ('acc, 'e) Base__.Result.t) -> ('acc, 'e) Base__.Result.t

val fold_until : t -> init:'acc -> f:('acc -> Elt.t -> ('acc, 'final) Base__Container_intf.Continue_or_stop.t) -> finish:('acc -> 'final) -> 'final

val exists : t -> f:(Elt.t -> bool) -> bool

val for_all : t -> f:(Elt.t -> bool) -> bool

val count : t -> f:(Elt.t -> bool) -> int

val sum : (module Base__Container_intf.Summable with type t = 'sum) -> t -> f:(Elt.t -> 'sum) -> 'sum

val find : t -> f:(Elt.t -> bool) -> Elt.t option

val find_map : t -> f:(Elt.t -> 'a option) -> 'a option

val to_list : t -> Elt.t list

val to_array : t -> Elt.t array

val min_elt : t -> compare:(Elt.t -> Elt.t -> int) -> Elt.t option

val max_elt : t -> compare:(Elt.t -> Elt.t -> int) -> Elt.t option