Module Basic0_container.On

On implements applicative traversal for a given applicative M.

Parameters

module M : Base.Applicative.S

Signature

include Generic_types.S0 with type t := t with type elt := Elt.t

include Basic_generic_on_applicative with type ('a, 'phantom) t := t and type 'a elt := Elt.t with module M := M

Generic refers to the container type as 'a t, and the element type as 'a elt; substitute t/elt (arity-0) or 'a t/'a (arity-1) accordingly below.

include Generic_types.Generic with type ('a, 'phantom) t := t with type 'a elt := Elt.t

val map_m : t -> f:(Elt.t -> Elt.t M.t) -> t M.t

map_m c ~f maps f over every t in c, threading through an applicative functor.

Example:

(* Travesty_base_exts.List adds applicative traversals to a list;
   With_errors (in S1_container) implements them on the On_error
   applicative functor. *)

let f x =
  Or_error.(if 0 < x then error_string "negative!" else ok x)
in
List.With_errors.map_m integers ~f