Module type Traversable_types.Basic0_on_applicative

Basic0_on_applicative is the inner signature of a traversal over arity-0 types.

include Generic_types.S0

type t

The container type.

type elt

The element type.

include Basic_generic_on_applicative with type ('a, 'phantom) t := t and type 'a elt := elt

Generic refers to the container type as 'a t, and the element type as 'a elt; substitute t/elt (arity-0) or 'a t/'a (arity-1) accordingly below.

include Generic_types.Generic with type ('a, 'phantom) t := t with type 'a elt := elt

module M : Base.Applicative.S

M is the applicative functor over which we're fold-mapping.

val map_m : t -> f:(elt -> elt M.t) -> t M.t

map_m c ~f maps f over every t in c, threading through an applicative functor.

Example:

(* Travesty_base_exts.List adds applicative traversals to a list;
   With_errors (in S1_container) implements them on the On_error
   applicative functor. *)

let f x =
  Or_error.(if 0 < x then error_string "negative!" else ok x)
in
List.With_errors.map_m integers ~f