Module type Traversable_types.Basic_generic_on_applicative

Basic_generic_on_applicative describes traversal on either an arity-0 or arity-1 type.

For arity-0 types, 'a t becomes t and 'a elt becomes elt.

For arity-1 types, 'a t becomes 'a t and 'a elt becomes 'a.

Generic refers to the container type as 'a t, and the element type as 'a elt; substitute t/elt (arity-0) or 'a t/'a (arity-1) accordingly below.

include Generic_types.Generic

type ('a, 'phantom) t

Placeholder for the container type.

type 'a elt

Placeholder for the type of elements inside the container.

module M : Base.Applicative.S

M is the applicative functor over which we're fold-mapping.

val map_m : ('a, 'phantom) t -> f:('a elt -> 'b elt M.t) -> ('b, 'phantom) t M.t

map_m c ~f maps f over every t in c, threading through an applicative functor.

Example:

(* Travesty_base_exts.List adds applicative traversals to a list;
   With_errors (in S1_container) implements them on the On_error
   applicative functor. *)

let f x =
  Or_error.(if 0 < x then error_string "negative!" else ok x)
in
List.With_errors.map_m integers ~f