Module type Traversable_types.Generic_on_monad

Generic_on_monad extends Generic_on_applicative to contain various derived operators that require monads; we use it to derive the signatures of the various On_monad modules.

For arity-0 types, 'a t becomes t and 'a elt becomes elt.

For arity-1 types, 'a t becomes 'a t and 'a elt becomes 'a.

module M : Base.Monad.S

The monad on which these operators are defined.

All applicative operators are available through lowering the monad to an applicative functor.

include Generic_on_applicative with module M := Monad_exts.App(M)

include Basic_generic_on_applicative with module M := Monad_exts.App(M)

Generic refers to the container type as 'a t, and the element type as 'a elt; substitute t/elt (arity-0) or 'a t/'a (arity-1) accordingly below.

include Generic_types.Generic

type 'a t

Placeholder for the container type.

type 'a elt

Placeholder for the type of elements inside the container.

val map_m : 'a t -> f:( 'a elt -> 'b elt Travesty.Monad_exts.App(M).t ) -> 'b t Travesty.Monad_exts.App(M).t

map_m c ~f maps f over every t in c, threading through an applicative functor.

Example:

(* Travesty_base_exts.List adds applicative traversals to a list;
   With_errors (in S1_container) implements them on the On_error
   applicative functor. *)

let f x =
  Or_error.(if 0 < x then error_string "negative!" else ok x)
in
List.With_errors.map_m integers ~f

val fold_map_m : 'a t -> f:( 'acc -> 'a elt -> ('acc * 'b elt) M.t ) -> init:'acc -> ('acc * 'b t) M.t

fold_map_m c ~f ~init folds f applicatively over every t in c, threading through an accumulator with initial value init.

val fold_m : 'a t -> init:'acc -> f:( 'acc -> 'a elt -> 'acc M.t ) -> 'acc M.t

fold_m x ~init ~f folds the applicative computation f over x, starting with initial value init, and returning the final value inside the applicative effect.

val iter_m : 'a t -> f:( 'a elt -> Base.unit M.t ) -> Base.unit M.t

iter_m x ~f iterates the computation f over x, returning the final applicative effect.

val mapi_m : f:( Base.int -> 'a elt -> 'b elt M.t ) -> 'a t -> 'b t M.t

mapi_m ~f x behaves as map_m, but also supplies f with the index of the element. This index should match the actual position of the element in the container x.