Module S0.With_errors

With_errors specialises On_monad to the error monad.

include Basic_generic_on_monad

Generic refers to the container type as 'a t, and the element type as 'a elt; substitute t/elt (arity-0) or 'a t/'a (arity-1) accordingly below.

include Generic_types.Generic

type 'a t

Placeholder for the container type.

type 'a elt

Placeholder for the type of elements inside the container.

module M : Base.Monad.S

M is the monad over which we're fold-mapping.

val map_m : 'a t -> f:('a elt -> 'b elt M.t) -> 'b t M.t

map_m c ~f maps f over every t in c, threading through monadic state.

Example:

(* Travesty_base_exts.List adds monadic traversals to a list;
   With_errors (in S1_container) implements them on the On_error
   monad. *)

let f x =
  Or_error.(if 0 < x then error_string "negative!" else ok x)
in
List.With_errors.map_m integers ~f

val fold_map_m : 'a t -> f:('acc -> 'a elt -> ('acc * 'b elt) M.t) -> init:'acc -> ('acc * 'b t) M.t

fold_map_m c ~f ~init folds f monadically over every t in c, threading through an accumulator with initial value init.

val fold_m : 'a t -> init:'acc -> f:('acc -> 'a elt -> 'acc M.t) -> 'acc M.t

fold_m x ~init ~f folds the monadic computation f over x, starting with initial value init, and returning the final value inside the monadic effect.

val iter_m : 'a t -> f:('a elt -> Base.unit M.t) -> Base.unit M.t

iter_m x ~f iterates the monadic computation f over x, returning the final monadic effect.

val mapi_m : f:(Base.int -> 'a elt -> 'b elt M.t) -> 'a t -> 'b t M.t

mapi_m ~f x behaves as mapM, but also supplies f with the index of the element. This index should match the actual position of the element in the container x.