Parameter Traverse0_right.2-S

include Travesty.Generic_types.Bi0

type t

Fixed type of containers.

type left

Fixed type of left elements.

type right

Fixed type of right elements.

include Travesty.Bi_traversable_types.Generic with type ('l, 'r) t := t and type 'l left := left and type 'r right := right

include Travesty.Generic_types.Bi_generic

type ('l, 'r) t

Generic container type.

type 'l left

Generic left type.

type 'r right

Generic right type.

We can do non-monadic bi-mapping operations.

include Travesty.Bi_mappable_types.Generic with type ('l, 'r) t := ('l, 'r) t and type 'l left := 'l left and type 'r right := 'r right

include Travesty.Bi_mappable_types.Basic_generic

include Travesty.Generic_types.Bi_generic

type ('l, 'r) t

Generic container type.

type 'l left

Generic left type.

type 'r right

Generic right type.

val bi_map : ('l1, 'r1) t -> left:('l1 left -> 'l2 left) -> right:('r1 right -> 'r2 right) -> ('l2, 'r2) t

bi_map c ~left ~right maps left over every 'l1 left, and right over every 'r1 right, in c.

val map_left : ('l1, 'r) t -> f:('l1 left -> 'l2 left) -> ('l2, 'r) t

map_left c ~f maps f over the left type of c only.

val map_right : ('l, 'r1) t -> f:('r1 right -> 'r2 right) -> ('l, 'r2) t

map_right c ~f maps f over the right type of c only.

module On_monad : functor (M : Base.Monad.S) -> Travesty.Bi_traversable_types.Generic_on_monad with type ('l, 'r) t := ('l, 'r) t and type 'l left := 'l left and type 'r right := 'r right and module M := M

On_monad implements monadic bi-traversal operators for a given monad M.

module With_errors : Travesty.Bi_traversable_types.Generic_on_monad with type ('l, 'r) t := ('l, 'r) t and type 'l left := 'l left and type 'r right := 'r right and module M := Base.Or_error

With_errors specialises On_monad to the error_monad.