Module Travesty__Filter_mappable

Simultaneous mapping and filtering.

Filter_mappable contains signatures and extensions for types that can be non-monadically mapped over, while simultaneously removing values.

Signatures

For input and output module signatures for this module's functors, see Filter_mappable_types.

Making full instances from basic ones

module Make0 : functor (F : Travesty.Filter_mappable_types.Basic0) -> Travesty.Filter_mappable_types.S0 with type t := F.t and type elt := F.elt

Make0 makes an S0 from a Basic0.

module Make1 : functor (F : Travesty.Filter_mappable_types.Basic1) -> Travesty.Filter_mappable_types.S1 with type 'a t := 'a F.t

Make0 makes an S1 from a Basic1.

Deriving Mappable

module To_mappable0 : functor (F : Travesty.Filter_mappable_types.Basic0) -> Travesty.Mappable_types.S0 with type t := F.t and type elt := F.elt

To_mappable0 lowers a filter-mappable arity-0 type to a mappable one.

module To_mappable1 : functor (F : Travesty.Filter_mappable_types.Basic1) -> Travesty.Mappable_types.S1 with type 'a t := 'a F.t

To_mappable1 lowers a filter-mappable arity-1 type to a mappable one.