Module Travesty__Mappable

Non-monadic mapping.

Mappable contains signatures and extensions for types that can be (non-monadically) mapped over. It resembles the Haskell (but not the OCaml!) notion of a functor, though we call the mapping function map.

Signatures

For input and output module signatures for this module's functors, see Mappable_types.

Extending mappable containers

We define several derived functions for mappable containers in Mappable_types---here, we define functors to generate them.

module Extend1 : functor (S : Travesty.Mappable_types.S1_container) -> Travesty.Mappable_types.Extensions1 with type 'a t := 'a S.t

Extend1 implements Extensions1 for an arity-1 mappable container.