Module Travesty__Monad_exts

Generic monad extensions.

Monad_exts contains a signature and functor for adding various extensions to Base-style monads.

Signatures

For input and output module signatures for this module's functors, see Filter_mappable_types.

Extensions functors

module Extend : functor (M : Base.Monad.S) -> Travesty.Monad_exts_types.S with type 'a t := 'a M.t

Extend creates extensions for a Monad.S.

Miscellaneous

module S2_to_S : functor (M : Base.Monad.S2) -> functor (B : Base.T) -> Base.Monad.S with type 'a S2_to_S.t := ('a, B.t) M.t

Demotes an arity-2 monad M to an arity-1 one, fixing its second type to be B.t.

module To_mappable : functor (M : Base.Monad.S) -> Travesty.Mappable_types.S1 with type 'a t := 'a M.t

Converts a monad to a mappable over M.map.