Module type Travesty__Monad_transform.S_fixed

S_fixed is the signature of monad transformers fixed to a particular inner monad.

type 'a t

The type of the outer transformer.

module Inner : Base.Monad.S

Inner is the inner monad.

val lift : 'a Inner.t -> 'a Inner.t t

lift x lifts x from the inner monad to the outer one.