Module type Travesty__State_transform_types.S2

S2 is the signature of state transformers parametrised over both value and state types.

include Base.Monad.S2

type ('a, 'e) t

val (>>=) : ('a, 'e) t -> ('a -> ('b, 'e) t) -> ('b, 'e) t

val (>>|) : ('a, 'e) t -> ('a -> 'b) -> ('b, 'e) t

module Let_syntax : sig ... end

module Monad_infix : sig ... end

val bind : ('a, 'e) t -> f:('a -> ('b, 'e) t) -> ('b, 'e) t

val return : 'a -> ('a, 'b) t

val map : ('a, 'e) t -> f:('a -> 'b) -> ('b, 'e) t

val join : (('a, 'e) t, 'e) t -> ('a, 'e) t

val ignore_m : ('a, 'e) t -> (unit, 'e) t

val all : ('a, 'e) t list -> ('a list, 'e) t

val all_unit : (unit, 'e) t list -> (unit, 'e) t

include Generic with type ('a, 's) t := ('a, 's) t and type 's state := 's

include Generic_builders with type 'a final := 'a

include Generic_types

type ('a, 's) t

t is the type of the state monad.

type 'a final

final is the type of returned results. In transformers, this becomes 'a Inner.t; otherwise, it becomes just 'a.

type 's state

state is the type used to represent the state outside of its monad. In S, 's state becomes x for some type x; in S2, 's state becomes 's.

val make : ('s state -> ('s state * 'a) final) -> ('a, 's) t

make creates a context-sensitive computation that can modify both the current context and the data passing through.

Specialised builders

val peek : ('s state -> 'a final) -> ('a, 's) t

peek creates a context-sensitive computation that can look at the current context, but not modify it.

val modify : ('s state -> 's state final) -> (Base.unit, 's) t

modify creates a context-sensitive computation that can look at and modify the current context.

val return : 'a final -> ('a, 's) t

return lifts a value or monad into a stateful computation.

module Inner : Base.Monad.S

Inner is the monad to which we're adding state.

State transformers have the same runner signatures as state monads, but lifted into the inner monad.

include Generic_runners with type ('a, 's) t := ('a, 's) t and type 'a final := 'a Inner.t and type 's state := 's state

include Generic_types

type ('a, 's) t

t is the type of the state monad.

type 'a final

final is the type of returned results. In transformers, this becomes 'a Inner.t; otherwise, it becomes just 'a.

type 's state

state is the type used to represent the state outside of its monad. In S, 's state becomes x for some type x; in S2, 's state becomes 's.

val run' : ('a, 's) t -> 's state -> ('s state * 'a) final

run' unfolds a t into a function from context to final state and result.

val run : ('a, 's) t -> 's state -> 'a final

run unfolds a t into a function from context to final result. To get the final context, use run' or call peek at the end of the computation.

include Fix with type ('a, 's) t := ('a, 's) t

type ('a, 's) t

val fix : f:(('a -> ('a, 's) t) -> 'a -> ('a, 's) t) -> 'a -> ('a, 's) t

fix ~f init builds a fixed point on f.

At each step, f is passed a continuation mu and a value a. It may choose to return a recursive application of mu, or some value derived from a.

To begin with, f is applied to mu and init.

module Monadic : Generic_builders with type 'a state := 'a state and type 'a final := 'a Inner.t and type ('a, 's) t := ('a, 's) t

Monadic contains a version of the builder interface that can interact with the inner monad (Inner) this state transformer is overlaying.