Module 1-Outer.On_monad

On_monad implements traversal operators for a given monad M. Compared to On(Monad_exts.App(M)), this adds various derived operators available only for monads.

Parameters

M : Base.Monad.S

Signature

module M : Base.Monad.S

The monad on which these operators are defined.

All applicative operators are available through lowering the monad to an applicative functor.

include Travesty.Traversable_types.Generic_on_applicative with module M := Travesty.Monad_exts.App(M)

include Travesty.Traversable_types.Basic_generic_on_applicative

Generic refers to the container type as 'a t, and the element type as 'a elt; substitute t/elt (arity-0) or 'a t/'a (arity-1) accordingly below.

include Travesty.Generic_types.Generic

type 'a t

Placeholder for the container type.

type 'a elt

Placeholder for the type of elements inside the container.

module M : Base.Applicative.S

M is the applicative functor over which we're fold-mapping.

val map_m : 'a t -> f:('a elt -> 'b elt M.t) -> 'b t M.t

map_m c ~f maps f over every t in c, threading through an applicative functor.

Example:

(* Travesty_base_exts.List adds applicative traversals to a list;
   With_errors (in S1_container) implements them on the On_error
   applicative functor. *)

let f x =
  Or_error.(if 0 < x then error_string "negative!" else ok x)
in
List.With_errors.map_m integers ~f

val iter_m : 'a t -> f:('a elt -> Base.unit M.t) -> Base.unit M.t

iter_m x ~f iterates the computation f over x, returning the final applicative effect.

val fold_map_m : 'a t -> f:('acc -> 'a elt -> ('acc * 'b elt) M.t) -> init:'acc -> ('acc * 'b t) M.t

fold_map_m c ~f ~init folds f applicatively over every t in c, threading through an accumulator with initial value init.

val fold_m : 'a t -> init:'acc -> f:('acc -> 'a elt -> 'acc M.t) -> 'acc M.t

fold_m x ~init ~f folds the applicative computation f over x, starting with initial value init, and returning the final value inside the applicative effect.

val iter_m : 'a t -> f:('a elt -> Base.unit M.t) -> Base.unit M.t

iter_m x ~f iterates the computation f over x, returning the final applicative effect.

val mapi_m : f:(Base.int -> 'a elt -> 'b elt M.t) -> 'a t -> 'b t M.t

mapi_m ~f x behaves as map_m, but also supplies f with the index of the element. This index should match the actual position of the element in the container x.