Module S.On_monad

On_monad provides various zipper operations parametrised by a monad.

Parameters

module M : Base.Monad.S

Signature

val pop_m : 'a t -> on_empty:( 'a t -> ('a * 'a t) M.t ) -> ('a * 'a t) M.t

pop_m zipper ~on_empty behaves like pop, but executes a custom monadic action on_empty, instead of returning an error, when the cursor is empty.

val peek_m : ?steps:Base.int -> 'a t -> on_empty:( 'a t -> 'a M.t ) -> 'a M.t

peek_m ?steps zipper ~on_empty behaves like peek_opt, but executes a custom monadic action on_empty, instead of returning None, when the cursor is empty.

val step_m : ?steps:Base.int -> 'a t -> on_empty:( 'a t -> 'a t M.t ) -> 'a t M.t

step_m ?steps zipper ~on_empty behaves like step, but executes a custom monadic action on_empty, instead of returning an error, when the cursor is empty.

val map_m_head : 'a t -> f:( 'a -> 'a Base.option M.t ) -> on_empty:( 'a t -> 'a t M.t ) -> 'a t M.t

map_m_head ?steps zipper ~on_empty behaves like map_head, but executes a custom monadic action on_empty, instead of leaving the zipper unchanged, when the cursor is empty.