Module type Bi_traversable_types.Basic1_right_on_applicative

Basic1_right_on_applicative is the inner signature of a monadic bi-traversal over arity-1 types with a floating right type and fixed left type.

include Generic_types.Bi_right

type 'r t

Partially fixed type of containers.

type left

Fixed type of left elements.

include Basic_generic_on_applicative with type ('l, 'r) t := 'r t and type 'l left := left and type 'r right := 'r

include Generic_types.Bi_generic

type ('l, 'r) t

Generic container type.

type 'l left

Generic left type.

type 'r right

Generic right type.

module M : Base.Applicative.S

M is the applicative functor over which we're bi-traversing.

val bi_map_m : ('l1, 'r1) t -> left:('l1 left -> 'l2 left M.t) -> right:('r1 right -> 'r2 right M.t) -> ('l2, 'r2) t M.t

bi_map_m c ~left ~right traverses with left over every 'l1 left, and right over every 'r1 right, in c.