Module Travesty.Container_exts

Signatures

For input and output module signatures for this module's functors, see Container_exts_types.

Extension functors

These functors extend Core containers with the extensions described in S0 and S1.

module Extend0 : functor (C : Base.Container.S0) -> Container_exts_types.S0 with type t := C.t and type elt := C.elt

Extend0 creates extensions for a Container.S0.

module Extend0_predicate : functor (P : Base.T) -> functor (C : Base.Container.S0 with type Extend0_predicate.elt = P.t -> Base.bool) -> Container_exts_types.S0_predicate with type t := C.t and type item := P.t

Extend0_predicate creates extensions for a Container.S0 over predicates.

module Extend1 : functor (C : Base.Container.S1) -> Container_exts_types.S1 with type 'a t := 'a C.t

Extend1 creates extensions for a Container.S1.