Module State.Make

Make makes an S (state monad with fixed state type) from a single state type.

Parameters

B : Base.T

Signature

type state = B.t

The fixed state type.

include Base.Monad.S

type 'a t

val (>>=) : 'a t -> ('a -> 'b t) -> 'b t

val (>>|) : 'a t -> ('a -> 'b) -> 'b t

module Monad_infix : sig ... end

val bind : 'a t -> f:('a -> 'b t) -> 'b t

val return : 'a -> 'a t

val map : 'a t -> f:('a -> 'b) -> 'b t

val join : 'a t t -> 'a t

val ignore_m : 'a t -> unit t

val all : 'a t list -> 'a list t

val all_unit : unit t list -> unit t

module Let_syntax : sig ... end

include Monad_exts_types.S with type 'a t := 'a t

type 'a t

The type of the extended monad.

S subsumes S_let.

include Monad_exts_types.S_let with type 'a t := 'a t

type 'a t

The type of the extended monad.

val let+ : 'a t -> ('a -> 'b) -> 'b t

let+ is map.

val let* : 'a t -> ('a -> 'b t) -> 'b t

let* is bind.

Haskell-style operators

val then_m : _ t -> 'a t -> 'a t

then_m x y sequentially composes the actions x and y as with >>=, but, rather than using the returned value of x, it instead just returns y.

val (>>) : _ t -> 'a t -> 'a t

x >> y is then_m x y.

val compose_m : ('a -> 'b t) -> ('b -> 'c t) -> 'a -> 'c t

compose_m f g is the Kleisli composition of f and g.

val (>=>) : ('a -> 'b t) -> ('b -> 'c t) -> 'a -> 'c t

x >=> y is compose_m x y.

Guarded monadic computations

val map_when_m : ?⁠otherwise:('a -> 'a t) -> bool -> 'a -> f:('a -> 'a t) -> 'a t

map_when_m ?otherwise condition ~f a is f a when condition is true, and otherwise a (by default, return) otherwise.

val when_m : ?⁠otherwise:(unit -> unit t) -> bool -> f:(unit -> unit t) -> unit t

when_m ?otherwise condition ~f is f () when condition is true, and otherwise () (by default, return) otherwise.

val map_unless_m : ?⁠otherwise:('a -> 'a t) -> bool -> 'a -> f:('a -> 'a t) -> 'a t

map_unless_m ?otherwise condition ~f a is f a when condition is false, and otherwise a (by default, return) otherwise.

val unless_m : ?⁠otherwise:(unit -> unit t) -> bool -> f:(unit -> unit t) -> unit t

unless_m ?otherwise condition ~f is f () when condition is false, and otherwise () (by default, return) otherwise.

Executing monadic effects in the middle of pipelines

val tee_m : 'a -> f:('a -> unit t) -> 'a t

tee_m val ~f executes f val for its monadic action, then returns val.

Example (using an extended Or_error):

let fail_if_negative x =
  On_error.when_m (Int.is_negative x) ~f:(fun () ->
      Or_error.error_string "value is negative!" )
in
Or_error.(42 |> tee_m ~f:fail_if_negative >>| fun x -> x * x)

(* Ok (1764) *)

val tee : 'a -> f:('a -> unit) -> 'a t

tee val ~f behaves as tee_m, but takes a non-monadic f.

Example (using an extended Or_error):

let print_if_negative x =
  if Int.negative x then Stdio.print_string "value is negative!"
in
Or_error.(
  try_get_value () >>= tee ~f:print_if_negative >>= try_use_value ())

include State_types.Generic with type ('a, 's) t := 'a t and type 's state := state and type 'a final := 'a

State monads share the signatures of their builder functions with state transformers...

include State_transform_types.Generic_builders

include State_transform_types.Generic_types

type ('a, 's) t

t is the type of the state monad.

type 'a final

final is the type of returned results. In transformers, this becomes 'a Inner.t; otherwise, it becomes just 'a.

type 's state

state is the type used to represent the state outside of its monad. In S, 's state becomes x for some type x; in S2, 's state becomes 's.

val make : ('s state -> ('s state * 'a) final) -> ('a, 's) t

make creates a context-sensitive computation that can modify both the current context and the data passing through.

Specialised builders

val peek : ('s state -> 'a final) -> ('a, 's) t

peek creates a context-sensitive computation that can look at the current context, but not modify it.

val modify : ('s state -> 's state final) -> (Base.unit, 's) t

modify creates a context-sensitive computation that can look at and modify the current context.

val return : 'a final -> ('a, 's) t

return lifts a value or monad into a stateful computation.

...as well as their runner functions...

include State_transform_types.Generic_runners with type ('a, 's) t := ('a, 's) t and type 'a final := 'a final and type 's state := 's state

include State_transform_types.Generic_types

type ('a, 's) t

t is the type of the state monad.

type 'a final

final is the type of returned results. In transformers, this becomes 'a Inner.t; otherwise, it becomes just 'a.

type 's state

state is the type used to represent the state outside of its monad. In S, 's state becomes x for some type x; in S2, 's state becomes 's.

val run' : ('a, 's) t -> 's state -> ('s state * 'a) final

run' unfolds a t into a function from context to final state and result.

val run : ('a, 's) t -> 's state -> 'a final

run unfolds a t into a function from context to final result. To get the final context, use run' or call peek at the end of the computation.

...and fixed-point combinators.

include State_transform_types.Fix with type ('a, 's) t := ('a, 's) t

type ('a, 's) t

val fix : f:(('a -> ('a, 's) t) -> 'a -> ('a, 's) t) -> 'a -> ('a, 's) t

fix ~f init builds a fixed point on f.

At each step, f is passed a continuation mu and a value a. It may choose to return a recursive application of mu, or some value derived from a.

To begin with, f is applied to mu and init.