Module Travesty_containers.Zipper

Signatures

For input and output signatures for this module's functors, see Zipper_types.

Plain zippers

module Plain : Zipper_types.S

Plain is a basic list zipper, without specialised functionality.

Marked zippers

module Make_marked : functor (Mark : Zipper_types.Basic_mark) -> Zipper_types.S_marked with type mark := Mark.t

Make_marked makes a marked zipper from a Basic_mark.

module Int_mark_zipper : Zipper_types.S_marked with type mark := Base.int

Int_mark_zipper is a marked zipper whose marks are integers.