Module type Zipper_types.S_marked_monadic

S_marked_monadic extends S_monadic to add functions for manipulating marks.

include S_monadic

type 'a t

module M : Base.Monad.S

val pop_m : 'a t -> on_empty:('a t -> ('a * 'a t) M.t) -> ('a * 'a t) M.t

pop_m zipper ~on_empty behaves like pop, but executes a custom monadic action on_empty, instead of returning an error, when the cursor is empty.

val peek_m : ?steps:Base.int -> 'a t -> on_empty:('a t -> 'a M.t) -> 'a M.t

peek_m ?steps zipper ~on_empty behaves like peek_opt, but executes a custom monadic action on_empty, instead of returning None, when the cursor is empty.

val step_m : ?steps:Base.int -> 'a t -> on_empty:('a t -> 'a t M.t) -> 'a t M.t

step_m ?steps zipper ~on_empty behaves like step, but executes a custom monadic action on_empty, instead of returning an error, when the cursor is empty.

val map_m_head : 'a t -> f:('a -> 'a Base.option M.t) -> on_empty:('a t -> 'a t M.t) -> 'a t M.t

map_m_head ?steps zipper ~on_empty behaves like map_head, but executes a custom monadic action on_empty, instead of leaving the zipper unchanged, when the cursor is empty.

type mark

The type of marks.

val mark_m : 'a t -> mark:mark -> on_empty:('a t -> 'a t M.t) -> 'a t M.t

mark_m zipper ~mark ~on_empty behaves like mark, but executes a custom monadic action on_empty, instead of returning an error, when the cursor is empty.

val recall_m : 'a t -> mark:mark -> on_empty:('a t -> 'a t M.t) -> 'a t M.t

recall_m zipper ~mark ~on_empty behaves like recall, but executes a custom monadic action on_empty, instead of returning an error, when the mark can't be found.

val delete_to_mark_m : 'a t -> mark:mark -> on_empty:('a t -> 'a t M.t) -> 'a t M.t

delete_to_mark_m zipper ~mark ~on_empty behaves like delete_to_mark, but executes a custom monadic action on_empty, instead of returning an error, when the mark can't be found.

val fold_m_until : 'a t -> f:('acc -> 'a -> 'a t -> (mark, 'a, 'acc, 'final) fold_outcome M.t) -> init:'acc -> finish:('acc -> 'a t -> 'final M.t) -> 'final M.t

fold_m_until zipper ~f ~init ~finish behaves like fold_until, except that f and finish, and therefore the function itself, return results inside a monad context.